Add handling for incomplete lvm/md devices. (#876441)
The strategy is to keep the vg/md in the devicetree, but to not include
them in devicetree.devices or the results of getDeviceBy{Name,Path}.
In the custom spoke we display a special options page for them with a
brief explanation and the option to remove the device or leave it be.

When removing an incomplete vg we don't actually do anything to remove
the vg since lvm gives us no way to specify one of several vgs with the
same name. All we do is wipe the pvs that we have access to.
(cherry picked from commit 531136b3eaf3dca07daa2f4a50fb8a5c980ef7fb)

+++ pyanaconda/storage/devices.py
@@ -2170,6 +2170,15 @@ class LVMVolumeGroupDevice(DMDevice):
     def _destroy(self):
         """ Destroy the device. """
         log_method_call(self, self.name, status=self.status)
+        if not self.complete:
+            for pv in self.pvs:
+                # Remove the PVs from the ignore filter so we can wipe them.
+                lvm.lvm_cc_removeFilterRejectRegexp(pv.name)
+
+            # Don't run vgremove or vgreduce since there may be another VG with
+            # the same name that we want to keep/use.
+            return
+
         lvm.vgreduce(self.name, [], rm=True)
         lvm.vgdeactivate(self.name)
         lvm.vgremove(self.name)
@@ -3001,6 +3010,10 @@ class MDRaidArrayDevice(StorageDevice):
                 rc = True
 
         return rc
+
+    @property
+    def complete(self):
+        return (self.memberDevices <= len(self.parents)) or not self.exists
 
     @property
     def devices(self):
